pyproject.toml
[tool.poetry]
name = "cadastre-query-service"
version = "0.1.0"
description = ""
authors = ["angelina <angln.pnn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
django = "4.2.10"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
redis = "^5.0.4"
celery = "^5.4.0"
requests = "^2.31.0"
django-celery-beat = "^2.6.0"
django-filter = "^24.2"
python-dotenv = "^1.0.1"
flake8 = "^7.0.0"
pytest = "^8.2.0"
pytest-django = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
